+++ packages/react/src/index.ts
@@ -2,6 +2,7 @@ export { BocetoView, type BocetoViewProps } from './BocetoView'
 export { BocetoEdit, type BocetoEditProps } from './BocetoEdit'
 export { BocetoPalette, type BocetoPaletteProps } from './BocetoPalette'
 export { BocetoInspector, type BocetoInspectorProps } from './BocetoInspector'
+export { BocetoComponents, type BocetoComponentsProps } from './BocetoComponents'
 export { BocetoEditFull, type BocetoEditFullProps } from './BocetoEditFull'
 
 // Side-effect-only import: pulls the `JSX.IntrinsicElements` augmentation