+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @boceto/mcp
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @boceto/edit@0.4.0
+
 ## 0.3.0
 
 ### Minor Changes